// src/common/enums/institutional-operational-state.enum.ts
export enum InstitutionalOperationalState {
  PENDING_PLANNING_INITIAL_VALIDATION = 'PENDING_PLANNING_INITIAL_VALIDATION',
  RETURNED_TO_PRODUCT_FROM_PLANNING = 'RETURNED_TO_PRODUCT_FROM_PLANNING',
  PENDING_FACTORY = 'PENDING_FACTORY',
  IN_FACTORY_PRODUCTION = 'IN_FACTORY_PRODUCTION',
  PENDING_PLANNING_PRODUCTION_VALIDATION = 'PENDING_PLANNING_PRODUCTION_VALIDATION',
  RETURNED_TO_FACTORY_FROM_PLANNING = 'RETURNED_TO_FACTORY_FROM_PLANNING',
  PENDING_LMS_UPLOAD = 'PENDING_LMS_UPLOAD',
  IN_LMS_UPLOAD = 'IN_LMS_UPLOAD',
  PENDING_PLANNING_LMS_VALIDATION = 'PENDING_PLANNING_LMS_VALIDATION',
  RETURNED_TO_LMS_FROM_PLANNING = 'RETURNED_TO_LMS_FROM_PLANNING',
  PENDING_PRODUCT_ACADEMIC_REVIEW = 'PENDING_PRODUCT_ACADEMIC_REVIEW',
  IN_PRODUCT_ACADEMIC_REVIEW = 'IN_PRODUCT_ACADEMIC_REVIEW',
  CHANGES_REQUESTED_BY_PRODUCT = 'CHANGES_REQUESTED_BY_PRODUCT',
  PENDING_PROJECT_RADICATION = 'PENDING_PROJECT_RADICATION',
  /** @deprecated Use PENDING_PROJECT_RADICATION */
  PENDING_FINAL_RADICATION = 'PENDING_PROJECT_RADICATION',
  FINALIZED = 'FINALIZED',
}